Subiendo versión nueva de ns-prueba

+++ src/modules/forms/services/form.service.ts
@@ -1,32 +1,63 @@
 // src/modules/forms/services/form.service.ts
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, InternalServerErrorException } from '@nestjs/common';
 import { CreateFormDto, UpdateFormDto } from '../dtos';
 import { FormRepository } from '../repositories/form.repository';
+import { Form, FormField, FormSection, ValidationRule } from '../entities';
 
 @Injectable()
 export class FormService {
   constructor(private readonly formRepository: FormRepository) {}
 
-  async findAll(page: number, limit: number) {
-    return this.formRepository.findAll(page, limit);
+  // Método para obtener todos los formularios
+  async findAll(page: number = 1, limit: number = 10) {
+    try {
+      return await this.formRepository.findAll(page, limit);
+    } catch (error) {
+      throw new InternalServerErrorException('Error al obtener formularios');
+    }
   }
 
+  // Método para obtener un formulario por ID
   async findById(id: string) {
     const form = await this.formRepository.findById(id);
     if (!form) throw new NotFoundException(`Formulario con ID ${id} no encontrado`);
     return form;
   }
 
+  // Método para crear un nuevo formulario
   async create(dto: CreateFormDto) {
-    // Enviar DTO completo (estructura JSON)
-    return await this.formRepository.createForm(dto);
+    //return await this.formRepository.createForm(dto);
+    try {
+      return await this.formRepository.createForm(dto);
+    } catch (error) {
+      throw new InternalServerErrorException('Error al crear el formulario');
+    }
   }
 
+  // Método para actualizar un formulario
   async update(id: string, dto: UpdateFormDto) {
-    return await this.formRepository.updateForm(id, dto);
+    const form = await this.formRepository.findById(id);
+    if (!form) throw new NotFoundException(`Formulario con ID ${id} no encontrado`);
+    
+    try {
+      return await this.formRepository.updateForm(id, dto);
+    } catch (error) {
+      throw new InternalServerErrorException('Error al actualizar el formulario');
+    }
   }
 
+
+
+
+  // Método para eliminar un formulario
   async remove(id: string): Promise<void> {
-    return await this.formRepository.removeForm(id);
+    const form = await this.formRepository.findById(id);
+    if (!form) throw new NotFoundException(`Formulario con ID ${id} no encontrado`);
+    
+    try {
+      await this.formRepository.removeForm(id);
+    } catch (error) {
+      throw new InternalServerErrorException('Error al eliminar el formulario');
+    }
   }
 }
